upd: push changes

+++ src/Architect/UI/EdgeInsets.cs
@@ -1,6 +1,6 @@
 namespace Architect.UI;
 
-public readonly record struct EdgeInsets(float Left, float Top, float Right, float Bottom)
+public readonly record struct EdgeInsets(int Left, int Top, int Right, int Bottom)
 {
     public readonly Size Size = new(Left + Right, Top + Bottom);
 
@@ -16,14 +16,14 @@ public readonly record struct EdgeInsets(float Left, float Top, float Right, flo
     public static EdgeInsets operator -(EdgeInsets a, EdgeInsets b) =>
         new(a.Left - b.Left, a.Top - b.Top, a.Right - b.Right, a.Bottom - b.Bottom);
 
-    public static EdgeInsets operator *(EdgeInsets a, float b) =>
+    public static EdgeInsets operator *(EdgeInsets a, int b) =>
         new(a.Left * b, a.Top * b, a.Right * b, a.Bottom * b);
 
-    public static EdgeInsets FromSide(float left = 0, float top = 0, float right = 0, float bottom = 0) =>
+    public static EdgeInsets FromSide(int left = 0, int top = 0, int right = 0, int bottom = 0) =>
         new(left, top, right, bottom);
 
-    public static EdgeInsets FromAxis(float horizontal = 0, float vertical = 0) =>
+    public static EdgeInsets FromAxis(int horizontal = 0, int vertical = 0) =>
         new(horizontal, vertical, horizontal, vertical);
 
-    public static EdgeInsets FromAll(float uniform) => new(uniform, uniform, uniform, uniform);
+    public static EdgeInsets FromAll(int uniform) => new(uniform, uniform, uniform, uniform);
 }
